fix: harden security, reduce complexity, and improve maintainability across codebase
- Harden against object-injection risks by removing dynamic object key access patterns
  and replacing with explicit allowlists and Map-backed storage.
- Remove regex backtracking in agentMode header parser to prevent ReDoS.
- Reduce any-type usage in mocks and tests.
- Extract large helper logic from provider.ts and extension.ts into dedicated
  modules (providerPromptUtils.ts, providerModelUtils.ts, participantOrchestration.ts)
  to lower file size and cyclomatic complexity.
- Update scripts/coverage-analysis.js to avoid dynamic object key lookups.
- Remove misleading hardcoded-password suppression in settings.ts.
